requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.28
python-dotenv==1.0.1
requests==2.31.0
discord.py==2.3.2
psycopg2-binary==2.9.9
gunicorn==21.2.0 